backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-filter
python-decouple
gunicorn
psycopg2-binary
dj-database-url
whitenoise

